# docs/swagger.yaml
basePath: /api
definitions:
  web_backend_internal_app_serializer.DepartmentApplicationDepartmentJSON:
    properties:
      department_application_id:
        type: integer
      department_id:
        type: integer
      direction:
        type: string
      main_department_id:
        type: integer
      role:
        type: string
      salary:
        type: number
      sort_order:
        type: integer
    type: object
  web_backend_internal_app_serializer.DepartmentApplicationJSON:
    properties:
      created_at:
        type: string
      creator_login:
        type: string
      department_application_id:
        type: integer
      finish_date:
        type: string
      forming_date:
        type: string
      incomplete_items_count:
        type: integer
      moderator_login:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
...
  web_backend_internal_app_serializer.SignInRequest:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  web_backend_internal_app_serializer.SignUpRequest:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  web_backend_internal_app_serializer.SignUpResponse:
    properties:
      is_moderator:
        type: boolean
      login:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://localhost:8080
  description: API для управления заявками отделов
  license:
    name: MIT
  title: Department Application API
  version: "1.0"
paths:
  /dep_app_dep/{department_id}/{department_application_id}:
    delete:
      description: Удаляет связь отдела и заявки
      parameters:
      - description: ID отдела
        in: path
        name: department_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: department_application_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная заявка
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationJSON'
        "400":
          description: Неверные ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Удалить отдел из заявки
      tags:
      - department_application_departments
    put:
      consumes:
      - application/json
      description: Обновляет параметры отдела в конкретной заявке
      parameters:
      - description: ID отдела
        in: path
        name: department_id
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: department_application_id
        required: true
        type: integer
      - description: Новые данные
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationDepartmentJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленные данные
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.DepartmentApplicationDepartmentJSON'
        "400":
          description: Неверные данные
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Не найдено
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Изменить данные отдела в заявке
      tags:
      - department_application_departments
...
  /users/signin:
    post:
      consumes:
      - application/json
      description: Принимает логин/пароль, возвращает jwt-токен в формате {"token":"..."}.
      parameters:
      - description: Логин и пароль
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: token
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Неверный запрос
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Пользователь не найден
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Вход (получение токена)
      tags:
      - users
  /users/signout:
    post:
      description: Удаляет токен текущего пользователя из blacklist. 204 No Content.
      produces:
      - application/json
      responses:
        "204":
          description: Токен добавлен в blacklist
        "400":
          description: Проблема с получением user_id
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка при удалении токена
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Выход (удаление токена)
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Возвращает login и is_moderator
        (без id и пароля).
      parameters:
      - description: Логин и пароль
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/web_backend_internal_app_serializer.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь создан
          schema:
            $ref: '#/definitions/web_backend_internal_app_serializer.SignUpResponse'
        "400":
          description: Ошибка валидации или входных данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"